Data engine config with live and backtest runners
